ci: run all tests across Vite 3, 6, 7, and 8-beta versions
Merged vite-compat tests into the main test suite under tests/compat/
and updated the CI workflow to run all unit, output, and e2e tests
with a Vite version matrix.

This ensures the plugin is properly tested against all supported Vite
versions as declared in peerDependencies.

Changes:
- Move tests/vite-compat/* to packages/vite-plugin/tests/compat/
- Add vite-version matrix to unit-output and e2e jobs
- Remove separate compat job (now redundant)
- Update test imports to use 'src' alias
