+++ packages/vite-plugin/tests/compat/.gitignore
@@ -1,3 +1,2 @@
 dist
 .chromium
-node_modules
