+++ packages/vite-plugin/tests/compat/e2e/src/content.js
