+++ packages/vite-plugin/tests/compat/plugins/src/content.js
