+++ packages/vite-plugin/tests/compat/vite-manifest/src/content.js
