--- tests/vite-compat/vitest.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-  test: {
-    testTimeout: 60000,
-    include: ['**/*.test.ts'],
-  },
-})
